Allow specifying output path for rasterio source `getRegion`
